service/requirements.txt
fastapi
webvtt-py
uvicorn